Renders relational links as arrays of links
When rendering JSON, we now always provide an _array_ of links for each
relation; this provides consistency to consumers, which now know they
will always receive an array, without needing to test for an array or
object.

XML rendering remains the same, as there is no way to represent a link
as an array; instead, you render multiple `<link>` elements with the
same `rel` attribute.